+++ ydb/core/tx/schemeshard/schemeshard__operation_alter_extsubdomain.cpp
@@ -2,6 +2,7 @@
 #include "schemeshard_impl.h"
 #include "schemeshard__operation_common.h"
 #include "schemeshard__operation_common_subdomain.h"
+#include "schemeshard_utils.h"  // for TransactionTemplate
 
 #include <ydb/core/base/subdomain.h>
 #include <ydb/core/base/hive.h>
