+++ ydb/core/tx/schemeshard/schemeshard__operation_alter_table.cpp
@@ -2,6 +2,8 @@
 #include "schemeshard__operation_common.h"
 #include "schemeshard_impl.h"
 
+#include "schemeshard_utils.h"  // for TransactionTemplate
+
 #include <ydb/core/base/subdomain.h>
 #include <ydb/core/base/hive.h>
