+++ ydb/core/tx/schemeshard/schemeshard__operation_create_cdc_stream.cpp
@@ -2,6 +2,9 @@
 
 #include "schemeshard__operation_part.h"
 #include "schemeshard__operation_common.h"
+
+#include "schemeshard_utils.h"  // for TransactionTemplate
+
 #include "schemeshard_cdc_stream_common.h"
 #include "schemeshard_impl.h"
