+++ ydb/core/tx/schemeshard/schemeshard__operation_create_table.cpp
@@ -3,6 +3,8 @@
 #include "schemeshard_impl.h"
 #include "schemeshard__op_traits.h"
 
+#include "schemeshard_utils.h"  // for IsAllowedKeyType
+
 #include <ydb/core/protos/flat_scheme_op.pb.h>
 #include <ydb/core/protos/datashard_config.pb.h>
