+++ ydb/core/tx/schemeshard/schemeshard__operation_move_index.cpp
@@ -4,6 +4,8 @@
 
 #include "schemeshard_impl.h"
 
+#include "schemeshard_utils.h"  // for TransactionTemplate
+
 #include <ydb/core/base/path.h>
 #include <ydb/core/mind/hive/hive.h>
 #include <ydb/core/protos/flat_tx_scheme.pb.h>
