+++ ydb/core/tx/schemeshard/schemeshard__table_stats.cpp
@@ -4,6 +4,7 @@
 #include <ydb/core/base/appdata.h>
 #include <ydb/core/base/cputime.h>
 #include <ydb/core/protos/sys_view.pb.h>
+#include <ydb/core/protos/table_stats.pb.h>
 
 namespace {
