+++ ydb/core/tx/schemeshard/schemeshard_utils.cpp
@@ -1,205 +1,12 @@
-#include "schemeshard_utils.h"
-
 #include <ydb/core/base/table_vector_index.h>
-#include <ydb/core/mind/hive/hive.h>
 #include <ydb/core/persqueue/utils.h>
-#include <ydb/core/protos/counters_schemeshard.pb.h>
+
+#include "schemeshard_info_types.h"
+
+#include "schemeshard_utils.h"
 
 namespace NKikimr {
 namespace NSchemeShard {
-
-void TShardDeleter::Shutdown(const NActors::TActorContext &ctx) {
-    for (auto& info : PerHiveDeletions) {
-        NTabletPipe::CloseClient(ctx, info.second.PipeToHive);
-    }
-    PerHiveDeletions.clear();
-}
-
-void TShardDeleter::SendDeleteRequests(TTabletId hiveTabletId,
-                                       const THashSet<TShardIdx> &shardsToDelete,
-                                       const THashMap<NKikimr::NSchemeShard::TShardIdx, NKikimr::NSchemeShard::TShardInfo>& shardsInfos,
-                                       const NActors::TActorContext &ctx) {
-    if (shardsToDelete.empty())
-        return;
-
-    TPerHiveDeletions& info = PerHiveDeletions[hiveTabletId];
-    if (!info.PipeToHive) {
-        NTabletPipe::TClientConfig clientConfig;
-        clientConfig.RetryPolicy = HivePipeRetryPolicy;
-        info.PipeToHive = ctx.Register(NTabletPipe::CreateClient(ctx.SelfID, ui64(hiveTabletId), clientConfig));
-    }
-    info.ShardsToDelete.insert(shardsToDelete.begin(), shardsToDelete.end());
-
-    for (auto shardIdx : shardsToDelete) {
-        ShardHive[shardIdx] = hiveTabletId;
-        // !HACK: use shardIdx as  TxId because Hive only replies with TxId
-        // TODO: change hive events to get rid of this hack
-        // svc@ in progress fixing it
-        TAutoPtr<TEvHive::TEvDeleteTablet> event = new TEvHive::TEvDeleteTablet(shardIdx.GetOwnerId(), ui64(shardIdx.GetLocalId()), ui64(shardIdx.GetLocalId()));
-        auto itShard = shardsInfos.find(shardIdx);
-        if (itShard != shardsInfos.end()) {
-            TTabletId shardTabletId = itShard->second.TabletID;
-            if (shardTabletId) {
-                event->Record.AddTabletID(ui64(shardTabletId));
-            }
-        }
-
-        Y_ABORT_UNLESS(shardIdx);
-
-        LOG_DEBUG_S(ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
-                    "Free shard " << shardIdx << " hive " << hiveTabletId << " at ss " << MyTabletID);
-
-        NTabletPipe::SendData(ctx, info.PipeToHive, event.Release());
-    }
-}
-
-void TShardDeleter::ResendDeleteRequests(TTabletId hiveTabletId, const THashMap<TShardIdx, TShardInfo>& shardsInfos, const NActors::TActorContext &ctx) {
-    LOG_NOTICE_S(ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
-                 "Resending tablet deletion requests from " << MyTabletID << " to " << hiveTabletId);
-
-    auto itPerHive = PerHiveDeletions.find(hiveTabletId);
-    if (itPerHive == PerHiveDeletions.end()) {
-        LOG_WARN_S(ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
-                   "Hive " << hiveTabletId << " not found for delete requests");
-        return;
-    }
-
-    THashSet<TShardIdx> toResend(std::move(itPerHive->second.ShardsToDelete));
-    PerHiveDeletions.erase(itPerHive);
-
-    SendDeleteRequests(hiveTabletId, toResend, shardsInfos, ctx);
-}
-
-void TShardDeleter::ResendDeleteRequest(TTabletId hiveTabletId,
-                                        const THashMap<TShardIdx, TShardInfo>& shardsInfos,
-                                        TShardIdx shardIdx,
-                                        const NActors::TActorContext &ctx) {
-    LOG_NOTICE_S(ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
-                 "Resending tablet deletion request from " << MyTabletID << " to " << hiveTabletId);
-
-    auto itPerHive = PerHiveDeletions.find(hiveTabletId);
-    if (itPerHive == PerHiveDeletions.end())
-        return;
-
-    auto itShardIdx = itPerHive->second.ShardsToDelete.find(shardIdx);
-    if (itShardIdx != itPerHive->second.ShardsToDelete.end()) {
-        THashSet<TShardIdx> toResend({shardIdx});
-        itPerHive->second.ShardsToDelete.erase(itShardIdx);
-        if (itPerHive->second.ShardsToDelete.empty()) {
-            PerHiveDeletions.erase(itPerHive);
-        }
-        SendDeleteRequests(hiveTabletId, toResend, shardsInfos, ctx);
-    } else {
-        LOG_WARN_S(ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
-                   "Shard " << shardIdx << " not found for delete request for Hive " << hiveTabletId);
-    }
-}
-
-void TShardDeleter::RedirectDeleteRequest(TTabletId hiveFromTabletId,
-                                          TTabletId hiveToTabletId,
-                                          TShardIdx shardIdx,
-                                          const THashMap<TShardIdx, TShardInfo>& shardsInfos,
-                                          const NActors::TActorContext &ctx) {
-    LOG_NOTICE_S(ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
-                 "Redirecting tablet deletion requests from " << hiveFromTabletId << " to " << hiveToTabletId);
-    auto itFromHive = PerHiveDeletions.find(hiveFromTabletId);
-    if (itFromHive != PerHiveDeletions.end()) {
-        auto& toHive(PerHiveDeletions[hiveToTabletId]);
-        auto itShardIdx = itFromHive->second.ShardsToDelete.find(shardIdx);
-        if (itShardIdx != itFromHive->second.ShardsToDelete.end()) {
-            toHive.ShardsToDelete.emplace(*itShardIdx);
-            itFromHive->second.ShardsToDelete.erase(itShardIdx);
-        } else {
-            LOG_WARN_S(ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
-                       "Shard " << shardIdx << " not found for delete request for Hive " << hiveFromTabletId);
-        }
-        if (itFromHive->second.ShardsToDelete.empty()) {
-            PerHiveDeletions.erase(itFromHive);
-        }
-    }
-
-    ResendDeleteRequest(hiveToTabletId, shardsInfos, shardIdx, ctx);
-}
-
-void TShardDeleter::ShardDeleted(TShardIdx shardIdx, const NActors::TActorContext &ctx) {
-    if (!ShardHive.contains(shardIdx))
-        return;
-
-    TTabletId hiveTabletId = ShardHive[shardIdx];
-    ShardHive.erase(shardIdx);
-    PerHiveDeletions[hiveTabletId].ShardsToDelete.erase(shardIdx);
-
-    if (PerHiveDeletions[hiveTabletId].ShardsToDelete.empty()) {
-        NTabletPipe::CloseClient(ctx, PerHiveDeletions[hiveTabletId].PipeToHive);
-        PerHiveDeletions.erase(hiveTabletId);
-    }
-}
-
-bool TShardDeleter::Has(TTabletId hiveTabletId, TActorId pipeClientActorId) const {
-    return PerHiveDeletions.contains(hiveTabletId) && PerHiveDeletions.at(hiveTabletId).PipeToHive == pipeClientActorId;
-}
-
-bool TShardDeleter::Has(TShardIdx shardIdx) const {
-    return ShardHive.contains(shardIdx);
-}
-
-bool TShardDeleter::Empty() const {
-    return PerHiveDeletions.empty();
-}
-
-void TSelfPinger::Handle(TEvSchemeShard::TEvMeasureSelfResponseTime::TPtr &ev, const NActors::TActorContext &ctx) {
-    Y_UNUSED(ev);
-    TInstant now = AppData(ctx)->TimeProvider->Now();
-    TDuration responseTime = now - SelfPingSentTime;
-    LastResponseTime = responseTime;
-    TabletCounters->Simple()[COUNTER_RESPONSE_TIME_USEC].Set(LastResponseTime.MicroSeconds());
-    if (responseTime.MilliSeconds() > 1000) {
-        LOG_WARN_S(ctx, NKikimrServices::FLAT_TX_SCHEMESHARD,
-                   "Schemeshard " << TabletId << " response time is " << responseTime.MilliSeconds() << " msec");
-    }
-    SelfPingInFlight = false;
-    if (responseTime > SELF_PING_INTERVAL) {
-        DoSelfPing(ctx);
-    } else {
-        ScheduleSelfPingWakeup(ctx);
-    }
-}
-
-void TSelfPinger::Handle(TEvSchemeShard::TEvWakeupToMeasureSelfResponseTime::TPtr &ev, const NActors::TActorContext &ctx) {
-    Y_UNUSED(ev);
-    SelfPingWakeupScheduled = false;
-    DoSelfPing(ctx);
-}
-
-void TSelfPinger::OnAnyEvent(const NActors::TActorContext &ctx) {
-    TInstant now = AppData(ctx)->TimeProvider->Now();
-    if (SelfPingInFlight) {
-        TDuration responseTime = now - SelfPingSentTime;
-        // Increase measured response time is ping is taking longer than then the previous one
-        LastResponseTime = Max(LastResponseTime, responseTime);
-        TabletCounters->Simple()[COUNTER_RESPONSE_TIME_USEC].Set(LastResponseTime.MicroSeconds());
-    } else if ((now - SelfPingWakeupScheduledTime) > SELF_PING_INTERVAL) {
-        DoSelfPing(ctx);
-    }
-}
-
-void TSelfPinger::DoSelfPing(const NActors::TActorContext &ctx) {
-    if (SelfPingInFlight)
-        return;
-
-    ctx.Send(ctx.SelfID, new TEvSchemeShard::TEvMeasureSelfResponseTime);
-    SelfPingSentTime = AppData(ctx)->TimeProvider->Now();
-    SelfPingInFlight = true;
-}
-
-void TSelfPinger::ScheduleSelfPingWakeup(const NActors::TActorContext &ctx) {
-    if (SelfPingWakeupScheduled)
-        return;
-
-    ctx.Schedule(SELF_PING_INTERVAL, new TEvSchemeShard::TEvWakeupToMeasureSelfResponseTime);
-    SelfPingWakeupScheduled = true;
-    SelfPingWakeupScheduledTime = AppData(ctx)->TimeProvider->Now();
-}
 
 PQGroupReserve::PQGroupReserve(const ::NKikimrPQ::TPQTabletConfig& tabletConfig, ui64 partitions) {
     Storage = partitions * NPQ::TopicPartitionReserveSize(tabletConfig);
@@ -555,7 +362,7 @@ bool ExtractTypes(const NKikimrSchemeOp::TTableDescription& baseTableDescr, TCol
 
         NScheme::TTypeInfo typeInfo;
         if (!GetTypeInfo(typeRegistry->GetType(typeName), column.GetTypeInfo(), typeName, columnName, typeInfo, explain)) {
-            return false; 
+            return false;
         }
 
         columnTypes[columnName] = typeInfo;
