+++ ydb/core/tx/schemeshard/schemeshard_utils.h
@@ -3,20 +3,13 @@
 #include "schemeshard_types.h"
 #include "schemeshard_info_types.h"
 
-#include <ydb/core/tablet/tablet_counters.h>
-#include <ydb/core/base/tablet_pipe.h>
 #include <ydb/core/base/table_index.h>
-#include <ydb/core/protos/table_stats.pb.h>
-
-#include <ydb/public/lib/scheme_types/scheme_type_id.h>
 
 #include <yql/essentials/minikql/mkql_type_ops.h>
 
-#include <ydb/library/actors/core/actorid.h>
-
 #include <util/generic/hash.h>
 #include <util/generic/hash_set.h>
-#include <util/generic/intrlist.h>
+
 
 namespace NKikimr {
 namespace NSchemeShard {
@@ -44,76 +37,6 @@ inline NKikimrSchemeOp::TModifyScheme TransactionTemplate(const TString& working
 
     return tx;
 }
-
-TSerializedCellVec ChooseSplitKeyByHistogram(const NKikimrTableStats::THistogram& histogram, ui64 total,
-                                  const TConstArrayRef<NScheme::TTypeInfo>& keyColumnTypes);
-
-class TShardDeleter {
-    struct TPerHiveDeletions {
-        TActorId PipeToHive;
-        THashSet<TShardIdx> ShardsToDelete;
-    };
-
-    ui64 MyTabletID;
-    // Hive TabletID -> non-acked deletions
-    THashMap<TTabletId, TPerHiveDeletions> PerHiveDeletions;
-    // Tablet -> Hive TabletID
-    THashMap<TShardIdx, TTabletId> ShardHive;
-    NTabletPipe::TClientRetryPolicy HivePipeRetryPolicy;
-
-public:
-    explicit TShardDeleter(ui64 myTabletId)
-        : MyTabletID(myTabletId)
-        , HivePipeRetryPolicy({})
-    {}
-
-    TShardDeleter(const TShardDeleter&) = delete;
-    TShardDeleter& operator=(const TShardDeleter&) = delete;
-
-    void Shutdown(const TActorContext& ctx);
-    void SendDeleteRequests(TTabletId hiveTabletId, const THashSet<TShardIdx>& shardsToDelete,
-                            const THashMap<TShardIdx, TShardInfo>& shardsInfos, const TActorContext& ctx);
-    void ResendDeleteRequests(TTabletId hiveTabletId,
-                              const THashMap<TShardIdx, TShardInfo>& shardsInfos, const TActorContext& ctx);
-    void ResendDeleteRequest(TTabletId hiveTabletId,
-                             const THashMap<TShardIdx, TShardInfo>& shardsInfos, TShardIdx shardIdx, const TActorContext& ctx);
-    void RedirectDeleteRequest(TTabletId hiveFromTabletId, TTabletId hiveToTabletId, TShardIdx shardIdx,
-                               const THashMap<TShardIdx, TShardInfo>& shardsInfos, const TActorContext& ctx);
-    void ShardDeleted(TShardIdx shardIdx, const TActorContext& ctx);
-    bool Has(TTabletId hiveTabletId, TActorId pipeClientActorId) const;
-    bool Has(TShardIdx shardIdx) const;
-    bool Empty() const;
-};
-
-// Self ping stuff
-class TSelfPinger {
-private:
-    static constexpr TDuration SELF_PING_INTERVAL = TDuration::MilliSeconds(1000);
-
-public:
-    TSelfPinger(TTabletId id, TTabletCountersBase* counters)
-        : TabletId(id)
-        , TabletCounters(counters)
-        , SelfPingInFlight(false)
-        , SelfPingWakeupScheduled(false)
-    {}
-
-    void Handle(TEvSchemeShard::TEvMeasureSelfResponseTime::TPtr &ev, const TActorContext &ctx);
-    void Handle(TEvSchemeShard::TEvWakeupToMeasureSelfResponseTime::TPtr &ev, const TActorContext &ctx);
-    void OnAnyEvent(const TActorContext &ctx);
-    void DoSelfPing(const TActorContext &ctx);
-    void ScheduleSelfPingWakeup(const TActorContext &ctx);
-
-private:
-    const TTabletId TabletId;
-    TTabletCountersBase * const TabletCounters;
-
-    TDuration LastResponseTime;
-    TInstant SelfPingSentTime;
-    bool SelfPingInFlight;
-    TInstant SelfPingWakeupScheduledTime;
-    bool SelfPingWakeupScheduled;
-};
 
 class PQGroupReserve {
 public:
