+++ ydb/core/tx/schemeshard/ut_bsvolume/ut_bsvolume.cpp
@@ -1,5 +1,6 @@
 #include <ydb/core/tx/schemeshard/ut_helpers/helpers.h>
-#include <ydb/core/tx/schemeshard/schemeshard_utils.h>
+
+#include <ydb/core/protos/blockstore_config.pb.h>
 
 using namespace NKikimr;
 using namespace NSchemeShard;
