+++ ydb/core/tx/schemeshard/ya.make
@@ -237,6 +237,10 @@ SRCS(
     schemeshard_path_element.cpp
     schemeshard_path_element.h
     schemeshard_schema.h
+    schemeshard_self_pinger.cpp
+    schemeshard_self_pinger.h
+    schemeshard_shard_deleter.cpp
+    schemeshard_shard_deleter.h
     schemeshard_svp_migration.cpp
     schemeshard_svp_migration.h
     schemeshard_tx_infly.h
